third_party/WebKit/LayoutTests/inspector/styles/pseudo-state-change-expected.txt

Tests pseudoStateChanged event dispatch.

Text

Running: dumpDefaultStyles
[expanded] 
element.style  { ()

[expanded] 
#inspected  { (pseudo-state-change.html:4 -> pseudo-state-change.html:4:1)
position: absolute;
top: 20px;
left: 0px;
color: black;
content: "Default";

[expanded] 
div  { (user agent stylesheet)
display: block;



Running: dumpHoveredStyles
[expanded] 
element.style  { ()

[expanded] 
#inspected:hover  { (pseudo-state-change.html:14 -> pseudo-state-change.html:14:1)
color: green;
content: "Hovered";

[expanded] 
#inspected  { (pseudo-state-change.html:4 -> pseudo-state-change.html:4:1)
position: absolute;
top: 20px;
left: 0px;
/-- overloaded --/ color: black;
/-- overloaded --/ content: "Default";

[expanded] 
div  { (user agent stylesheet)
display: block;



Running: dumpActiveStyles
[expanded] 
element.style  { ()

[expanded] 
#inspected:active  { (pseudo-state-change.html:24 -> pseudo-state-change.html:24:1)
color: red;
content: "Active";

[expanded] 
#inspected:focus  { (pseudo-state-change.html:19 -> pseudo-state-change.html:19:1)
/-- overloaded --/ color: blue;
/-- overloaded --/ content: "Focused";

[expanded] 
#inspected:hover  { (pseudo-state-change.html:14 -> pseudo-state-change.html:14:1)
/-- overloaded --/ color: green;
/-- overloaded --/ content: "Hovered";

[expanded] 
#inspected  { (pseudo-state-change.html:4 -> pseudo-state-change.html:4:1)
position: absolute;
top: 20px;
left: 0px;
/-- overloaded --/ color: black;
/-- overloaded --/ content: "Default";

[expanded] 
:focus  { (user agent stylesheet)
outline: -webkit-focus-ring-color auto 5px;

[expanded] 
div  { (user agent stylesheet)
display: block;



Running: dumpFocusedStyles
[expanded] 
element.style  { ()

[expanded] 
#inspected:focus  { (pseudo-state-change.html:19 -> pseudo-state-change.html:19:1)
color: blue;
content: "Focused";

[expanded] 
#inspected:hover  { (pseudo-state-change.html:14 -> pseudo-state-change.html:14:1)
/-- overloaded --/ color: green;
/-- overloaded --/ content: "Hovered";

[expanded] 
#inspected  { (pseudo-state-change.html:4 -> pseudo-state-change.html:4:1)
position: absolute;
top: 20px;
left: 0px;
/-- overloaded --/ color: black;
/-- overloaded --/ content: "Default";

[expanded] 
:focus  { (user agent stylesheet)
outline: -webkit-focus-ring-color auto 5px;

[expanded] 
div  { (user agent stylesheet)
display: block;



Running: dumpUnfocusedStyles
[expanded] 
element.style  { ()

[expanded] 
#inspected:hover  { (pseudo-state-change.html:14 -> pseudo-state-change.html:14:1)
color: green;
content: "Hovered";

[expanded] 
#inspected  { (pseudo-state-change.html:4 -> pseudo-state-change.html:4:1)
position: absolute;
top: 20px;
left: 0px;
/-- overloaded --/ color: black;
/-- overloaded --/ content: "Default";

[expanded] 
div  { (user agent stylesheet)
display: block;



Running: dumpUnhoveredStyles
[expanded] 
element.style  { ()

[expanded] 
#inspected  { (pseudo-state-change.html:4 -> pseudo-state-change.html:4:1)
position: absolute;
top: 20px;
left: 0px;
color: black;
content: "Default";

[expanded] 
div  { (user agent stylesheet)
display: block;
